+++ R/data.R
@@ -2,7 +2,7 @@
 #' @aliases cancer
 #' @description Cancer survival rates for 5, 10, 15, and 20 years for twenty-four cancer sites.
 #' @format
-#' A dataframe containing twenty-four observations for four years.
+#' A data frame containing twenty-four observations for four years.
 #' 
 #' @source Edward Tufte. \emph{Beautiful Evidence}. Graphics Press, 174-176.
 #' @keywords datasets
@@ -12,7 +12,7 @@
 #' @aliases gdp
 #' @description Current receipts of fifteen national governments as a percentage of gross domestic product
 #' @format
-#' A dataframe containing fifteen country observations for two years.
+#' A data frame containing fifteen country observations for two years.
 #' 
 #' @source Edward Tufte. \emph{Beautiful Evidence}. Graphics Press, 174-176.
 #' @keywords datasets
@@ -22,8 +22,18 @@
 #' @aliases states
 #' @description The relative rank of U.S. state populations for nine decades (1790-1870).
 #' @format
-#' A dataframe containing twenty-four observations for four years.
+#' A data frame containing twenty-four observations for four years.
 #' 
 #' @source Gray, Ormando Willis, and Colton, G. Woolworth. 1878. ``Chart Exhibiting the Relative Rank of the States for Nine Decades (1790-1870).'' Philadelphia: O.W. Gray & Son. Dataset constructed from slopegraph shown here: \url{http://www.davidrumsey.com/luna/servlet/detail/RUMSEY~8~1~207741~3003452:Chart-Exhibiting-the-Relative-Rank-}
 #' @keywords datasets
 "states"
+
+#' @title US Bachelors Degrees
+#' @aliases bachelors
+#' @description Bachelor's degrees conferred by postsecondary institutions, by field of study: Selected years, 1970-71 through 2013-14
+#' @format
+#' A data frame containing thirty-three observations for nineteen years.
+#' 
+#' @source  U.S. Department of Education, National Center for Education Statistics, Higher Education General Information Survey (HEGIS), \dQuote{Degrees and Other Formal Awards Conferred} surveys, 1970-71 through 1985-86; Integrated Postsecondary Education Data System (IPEDS), \dQuote{Completions Survey} (IPEDS-C:91-99); and IPEDS Fall 2000 through Fall 2014, Completions component. Retrieved 2016-12-08 from \url{http://nces.ed.gov/programs/digest/d15/tables/dt15_322.10.asp}.
+#' @keywords datasets
+"bachelors"